delete cmu nav
