--- data/.lfs/cmu_nav_paths.tar.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f88fdde4e1e6b5c0d0612c7e7cbf920d75805d471937952f195cb15b0543f037
-size 1291318
