--- data/.lfs/og_nav_60s.npz.tar.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a574137a3d661d36d2c2f07be2a1a56a1c75e2f9d3282ee32374e125af66d6e1
-size 108237809
