--- data/.lfs/unitree_g1_local_planner_precomputed_paths.tar.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:f8f5d128754569aecfec8c45fb1dfc158e2a773974fa0a4221c141925b964ff4
-size 1288141
