+++ dimos/hardware/drive_trains/flowbase/README.md
@@ -7,19 +7,14 @@ driver runs on the i2rt base.
 
 ```mermaid
 flowchart LR
-    MID["Livox MID-360<br/>(off-NUC)"] -->|Ethernet / point cloud| NUC
     subgraph NUC["Dev NUC"]
-        direction TB
-        F["FastLio2"] --> NS["nav stack"]
-        NS --> CC["ControlCoordinator"]
-        F --> R["Rerun viewer"]
-        NS --> R
+        CC["ControlCoordinator"]
     end
     NUC <-->|"Portal RPC @ 172.6.2.20:11323<br/>set_target_velocity / get_odometry"| PI["FlowBase Pi<br/>(driver as systemd, off-NUC)"]
 ```
 
-Sensor and base driver are both off-NUC. The dev NUC runs everything in
-between (FastLio2, nav stack, ControlCoordinator, Rerun viewer).
+The base driver is off-NUC. The dev NUC runs the ControlCoordinator, which
+speaks Portal RPC to the Pi.
 
 ## 1. FlowBase driver
 
@@ -48,7 +43,7 @@ nc -vz 172.6.2.20 11323
 
 ## 2. Launch
 
-Three blueprints, in increasing order of capability:
+Two blueprints:
 
 ```bash
 # Coordinator only (drive /cmd_vel from another source)
@@ -56,51 +51,18 @@ dimos run coordinator-flowbase
 
 # Coordinator + WASD pygame teleop
 dimos run coordinator-flowbase-keyboard-teleop
-
-# Coordinator + FastLio2 + nav stack + click-to-drive in Rerun
-LIDAR_HOST_IP=192.168.1.5 LIDAR_IP=192.168.1.189 \
-WAYLAND_DISPLAY=wayland-0 XDG_SESSION_TYPE=wayland \
-  dimos run coordinator-flowbase-nav
 ```
 
-All three use the `flowbase` adapter against `172.6.2.20:11323` and
+Both use the `flowbase` adapter against `172.6.2.20:11323` and
 publish/subscribe on LCM `/cmd_vel` + `/coordinator_joint_state`.
 
 ### Blueprint notes
 
-- **`coordinator-flowbase-nav`** composes `FastLio2` SLAM +
-  `create_cmu_nav(planner="simple")` + `MovementManager` (for click
-  forwarding and tele/nav velocity mux) + `ControlCoordinator` with the
-  FlowBase adapter as the driver. Mirrors `unitree-g1-nav-onboard` for the
-  perception/planning half. Click anywhere on the floor in the Rerun 3D
-  viewer → robot navigates there.
 - **`coordinator-flowbase-keyboard-teleop`** opens a small pygame window —
   **focus that window** to drive. Controls: W/S forward-back · Q/E strafe ·
   A/D turn · Shift boost · Ctrl slow · Space stop · ESC quit.
 - **`coordinator-flowbase`** is just the bare driver; you publish `/cmd_vel`
   from somewhere else.
-
-### Env defaults (nav blueprint)
-
-| Variable | Default | Meaning |
-|---|---|---|
-| `LIDAR_HOST_IP` | `192.168.1.5` | This machine's IP on the LIDAR subnet |
-| `LIDAR_IP` | `192.168.1.189` | Livox MID-360 IP |
-| `WAYLAND_DISPLAY` | (unset) | Set to `wayland-0` so the native Rerun viewer can render |
-| `XDG_SESSION_TYPE` | `tty` (in remote terminals) | Set to `wayland` for the same reason |
-
-Sensor mount is hardcoded in `_flowbase_mid360_mount` at 20cm forward /
-20cm right / 10cm up of base center, level orientation. Edit
-[mobile.py](../../../control/blueprints/mobile.py) to change.
-
-> **Sanity check before the nav blueprint**: run `dimos run mid360-fastlio`
-> first to verify the LIDAR locks in. If FastLio2 can't see the sensor,
-> nav won't start.
-
-> **`nix develop` breaks the Rerun viewer** — its `libvulkan` shadows the
-> system one and the native viewer fails surface creation. Run `dimos run`
-> from a plain `.venv` shell (native modules are cached after the first
-> build, so `nix develop` isn't needed at runtime).
 
 ## Notes
 
@@ -109,8 +71,3 @@ Sensor mount is hardcoded in `_flowbase_mid360_mount` at 20cm forward /
   ROS frame.
 - Address override: edit `_flowbase_twist_base(address=...)` in
   [mobile.py](../../../control/blueprints/mobile.py).
-- LIDAR mount: edit `_flowbase_mid360_mount` in the same file (set a
-  tilted quaternion if the MID-360 isn't level).
-- `local_planner` reuses `G1_LOCAL_PLANNER_PRECOMPUTED_PATHS` for now. If
-  local planning is unstable, generate FlowBase-specific paths as a
-  follow-up.
