--- dimos/navigation/cmu_nav/.gitignore
@@ -1,2 +1,0 @@
-# Cloned C++ source repos
-repo/
