--- dimos/navigation/cmu_nav/modules/path_follower/test_path_follower_rosbag.py
@@ -1,269 +1,0 @@
-# Copyright 2026 Dimensional Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-from pathlib import Path
-import sys
-import threading
-import time
-
-import lcm as lcmlib
-import numpy as np
-import pytest
-
-from dimos.constants import DEFAULT_THREAD_JOIN_TIMEOUT
-from dimos.msgs.geometry_msgs.Twist import Twist
-from dimos.navigation.cmu_nav.tests.deprecation import SKIP_CMU_NAV
-from dimos.navigation.cmu_nav.tests.rosbag_fixtures import (
-    LcmCollector,
-    NativeProcessRunner,
-    feed_at_original_timing,
-    lcm_handle_loop,
-    load_rosbag_window,
-)
-from dimos.utils.logging_config import setup_logger
-
-logger = setup_logger()
-
-pytestmark = [pytest.mark.self_hosted, SKIP_CMU_NAV]
-
-_PROCESS_STARTUP_SEC = 1.0
-_POST_FEED_DRAIN_SEC = 2.0
-
-PATH_FOLLOWER_BIN = Path(__file__).parent / "result" / "bin" / "path_follower"
-
-# LCM topics
-PATH_LCM = "/rbpf_path#nav_msgs.Path"
-ODOM_LCM = "/rbpf_odom#nav_msgs.Odometry"
-CMD_VEL_LCM = "/rbpf_cmd#geometry_msgs.Twist"
-SLOW_DOWN_LCM = "/rbpf_slow#std_msgs.Int8"
-SAFETY_STOP_LCM = "/rbpf_safety#std_msgs.Int8"
-
-# OG nav stack G1 config values
-OG_PATHFOLLOWER_ARGS = [
-    "--lookAheadDis",
-    "0.5",
-    "--maxSpeed",
-    "0.75",
-    "--autonomySpeed",
-    "0.75",
-    "--maxAccel",
-    "1.5",
-    "--maxYawRate",
-    "40.0",
-    "--yawRateGain",
-    "1.5",
-    "--stopYawRateGain",
-    "1.5",
-    "--goalYawGain",
-    "2.0",
-    "--slowDwnDisThre",
-    "0.875",
-    "--dirDiffThre",
-    "0.4",
-    "--stopDisThre",
-    "0.4",
-    "--omniDirGoalThre",
-    "0.5",
-    "--omniDirDiffThre",
-    "1.5",
-    "--twoWayDrive",
-    "false",  # OG runtime value (not omniDir default)
-    "--switchTimeThre",
-    "1.0",
-    "--autonomyMode",
-    "true",  # Set true at runtime by cross_wall_test.py
-    "--pubSkipNum",
-    "1",
-    "--noRotAtGoal",
-    "false",  # OG default
-    "--noRotAtStop",
-    "false",  # OG default
-    "--slowRate1",
-    "0.25",
-    "--slowRate2",
-    "0.5",
-    "--slowRate3",
-    "0.75",
-    "--slowTime1",
-    "2.0",
-    "--slowTime2",
-    "2.0",
-]
-
-
-class TestPathFollowerRosbag:
-    """Validate PathFollower accuracy against OG nav stack recording."""
-
-    def test_cmd_vel_accuracy(self) -> None:
-        """Feed path + odom at original timing and compare cmd_vel."""
-        if not PATH_FOLLOWER_BIN.exists():
-            pytest.skip(f"PathFollower binary not found: {PATH_FOLLOWER_BIN}")
-
-        window = load_rosbag_window()
-        ref_cmd = window.cmd_vel
-        assert len(ref_cmd) > 0, "No reference cmd_vel in fixture"
-
-        lcm = lcmlib.LCM()
-        cmd_collector = LcmCollector(topic=CMD_VEL_LCM, msg_type=Twist)
-        cmd_collector.start(lcm)
-
-        stop_event = threading.Event()
-        handle_thread = threading.Thread(
-            target=lcm_handle_loop, args=(lcm, stop_event), daemon=True
-        )
-        handle_thread.start()
-
-        runner = NativeProcessRunner(
-            binary_path=str(PATH_FOLLOWER_BIN),
-            args=[
-                "--path",
-                PATH_LCM,
-                "--odometry",
-                ODOM_LCM,
-                "--slow_down",
-                SLOW_DOWN_LCM,
-                "--safety_stop",
-                SAFETY_STOP_LCM,
-                "--cmd_vel",
-                CMD_VEL_LCM,
-                *OG_PATHFOLLOWER_ARGS,
-            ],
-        )
-
-        try:
-            runner.start()
-            assert runner.is_running, "PathFollower binary failed to start"
-            time.sleep(_PROCESS_STARTUP_SEC)
-
-            # Feed path + odom from the rosbag at original timing.
-            # PathFollower subscribes to /path (LocalPlanner output) and /odometry.
-            feed_at_original_timing(
-                lcm,
-                window,
-                topic_map={
-                    "odom": ODOM_LCM,
-                    "path": PATH_LCM,
-                },
-                odom_subsample=1,
-            )
-
-            time.sleep(_POST_FEED_DRAIN_SEC)
-
-        finally:
-            runner.stop()
-            stop_event.set()
-            handle_thread.join(timeout=DEFAULT_THREAD_JOIN_TIMEOUT)
-            cmd_collector.stop(lcm)
-
-        our_cmds = [(msg.linear.x, msg.linear.y, msg.angular.z) for msg in cmd_collector.messages]
-
-        ref_nonzero = ref_cmd[np.abs(ref_cmd[:, 1]) > 0.01]
-        our_nonzero = [c for c in our_cmds if abs(c[0]) > 0.01 or abs(c[1]) > 0.01]
-
-        ref_mean_speed = (
-            float(np.sqrt(ref_nonzero[:, 1] ** 2 + ref_nonzero[:, 2] ** 2).mean())
-            if len(ref_nonzero) > 0
-            else 0.0
-        )
-        our_mean_speed = (
-            float(np.mean([np.sqrt(lx**2 + ly**2) for lx, ly, _ in our_nonzero]))
-            if our_nonzero
-            else 0.0
-        )
-        speed_ratio = our_mean_speed / ref_mean_speed if ref_mean_speed > 0 else 0.0
-
-        # Speed comparison at multiple levels
-        ref_speeds = (
-            np.sqrt(ref_nonzero[:, 1] ** 2 + ref_nonzero[:, 2] ** 2)
-            if len(ref_nonzero) > 0
-            else np.array([])
-        )
-        our_speeds = (
-            np.array([np.sqrt(lx**2 + ly**2) for lx, ly, _ in our_nonzero])
-            if our_nonzero
-            else np.array([])
-        )
-
-        # Steady-state comparison: filter to speeds > 0.5 m/s (fully in autonomy,
-        # past acceleration ramp, not in joy-gated zero phase)
-        ref_steady = ref_speeds[ref_speeds > 0.5] if len(ref_speeds) > 0 else np.array([])
-        our_steady = our_speeds[our_speeds > 0.5] if len(our_speeds) > 0 else np.array([])
-        steady_ratio = (
-            float(our_steady.mean() / ref_steady.mean())
-            if len(ref_steady) > 0 and len(our_steady) > 0
-            else 0.0
-        )
-
-        logger.info(f"\n{'=' * 60}")
-        logger.info("PATH FOLLOWER DEVIATION SCORE")
-        logger.info(f"  Our cmd_vel:        {len(our_cmds)}")
-        logger.info(f"  Reference:          {len(ref_cmd)}")
-        logger.info(f"  Count ratio:        {len(our_cmds) / len(ref_cmd):.3f}")
-        logger.info(f"  Our non-zero:       {len(our_nonzero)}")
-        logger.info(f"  Ref non-zero:       {len(ref_nonzero)}")
-        logger.info(f"  Our mean speed:     {our_mean_speed:.3f} m/s")
-        logger.info(f"  Ref mean speed:     {ref_mean_speed:.3f} m/s")
-        logger.info(f"  Speed ratio (all):  {speed_ratio:.3f}")
-        logger.info(f"  Steady-state ratio: {steady_ratio:.3f}  (>0.5 m/s only)")
-        if len(our_speeds) > 0:
-            logger.info(f"  Our max speed:      {our_speeds.max():.3f} m/s")
-        if len(ref_speeds) > 0:
-            logger.info(f"  Ref max speed:      {ref_speeds.max():.3f} m/s")
-
-        # Yaw rate comparison (steady-state)
-        ref_yaws = np.abs(ref_nonzero[:, -1]) if len(ref_nonzero) > 0 else np.array([])
-        our_yaws = np.array([abs(az) for _, _, az in our_nonzero]) if our_nonzero else np.array([])
-        ref_steady_yaw = ref_yaws[ref_speeds > 0.5] if len(ref_yaws) > 0 else np.array([])
-        our_steady_yaw = our_yaws[our_speeds > 0.5] if len(our_yaws) > 0 else np.array([])
-        yaw_ratio = (
-            float(our_steady_yaw.mean() / ref_steady_yaw.mean())
-            if len(ref_steady_yaw) > 0 and len(our_steady_yaw) > 0 and ref_steady_yaw.mean() > 0.01
-            else 1.0  # If ref yaw is near-zero, skip ratio check
-        )
-
-        logger.info(f"  Yaw rate ratio:     {yaw_ratio:.3f}  (steady-state)")
-        logger.info(f"{'=' * 60}\n")
-
-        # Assertions (tightened to observed behavior ±5%)
-        assert len(our_cmds) > 0, "PathFollower produced no cmd_vel"
-        assert len(our_nonzero) > 0, "All cmd_vel are zero"
-
-        # Count ratio: we expect ~1.02x reference (timing jitter). macOS runs
-        # the binary's loop at a slower rate than Linux, so widen the lower
-        # bound there — correctness is enforced by the speed/yaw assertions.
-        count_ratio = len(our_cmds) / len(ref_cmd)
-        count_lower = 0.7 if sys.platform == "darwin" else 0.9
-        assert count_lower < count_ratio < 1.1, (
-            f"Message count ratio {count_ratio:.3f} outside [{count_lower}, 1.1]"
-        )
-
-        # Steady-state speed: observed 0.955, allow ±5%
-        assert 0.9 < steady_ratio < 1.05, (
-            f"Steady-state speed ratio {steady_ratio:.3f} outside [0.9, 1.05]"
-        )
-
-        # Max speed must match exactly (same autonomySpeed cap)
-        if len(ref_speeds) > 0 and len(our_speeds) > 0:
-            max_speed_ratio = float(our_speeds.max() / ref_speeds.max())
-            assert 0.95 < max_speed_ratio < 1.05, (
-                f"Max speed ratio {max_speed_ratio:.3f} outside [0.95, 1.05]"
-            )
-
-        # Yaw rate: should be in same ballpark (±30% — more variable than speed)
-        if ref_steady_yaw.mean() > 0.01:
-            assert 0.7 < yaw_ratio < 1.3, (
-                f"Steady-state yaw ratio {yaw_ratio:.3f} outside [0.7, 1.3]"
-            )
