--- dimos/robot/diy/alfred/blueprints/alfred_nav.py
@@ -1,79 +1,0 @@
-# Copyright 2025-2026 Dimensional Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-import os
-from typing import Any
-
-from dimos.core.coordination.blueprints import autoconnect
-from dimos.core.global_config import global_config
-from dimos.hardware.sensors.lidar.fastlio2.module import FastLio2
-from dimos.navigation.cmu_nav.main import cmu_nav_rerun_config, create_cmu_nav
-from dimos.navigation.movement_manager.movement_manager import MovementManager
-from dimos.robot.diy.alfred.config import LOCAL_PLANNER_PRECOMPUTED_PATHS
-from dimos.robot.diy.alfred.effector_high_level import AlfredHighLevel
-from dimos.visualization.vis_module import vis_module
-
-nav_config: dict[str, Any] = dict(
-    planner="simple",
-    vehicle_height=0.5,
-    max_speed=0.8,
-    terrain_analysis={
-        "obstacle_height_threshold": 0.15,
-        "ground_height_threshold": 0.10,
-        "sensor_range": 20,
-    },
-    local_planner={
-        "paths_dir": str(LOCAL_PLANNER_PRECOMPUTED_PATHS),
-        "publish_free_paths": False,
-    },
-    simple_planner={
-        "body_frame": "mid360_link",
-        "cell_size": 0.2,
-        "obstacle_height_threshold": 0.15,
-        "inflation_radius": 0.3,
-        "lookahead_distance": 2.0,
-        "replan_rate": 5.0,
-        "replan_cooldown": 2.0,
-    },
-)
-
-alfred_nav = (
-    autoconnect(
-        FastLio2.blueprint(
-            host_ip=os.getenv("LIDAR_HOST_IP", "192.168.1.5"),
-            lidar_ip=os.getenv("LIDAR_IP", "192.168.1.189"),
-        ),
-        create_cmu_nav(**nav_config),
-        MovementManager.blueprint(),
-        AlfredHighLevel.blueprint(),
-        vis_module(
-            global_config.viewer,
-            rerun_config={
-                **cmu_nav_rerun_config({"memory_limit": "1GB"}, vis_throttle=0.5),
-                "rerun_open": "native",
-            },
-        ),
-    )
-    .remappings(
-        [
-            # nav stack needs "registered_scan"
-            (FastLio2, "lidar", "registered_scan"),
-            # SimplePlanner / FarPlanner owns way_point — disconnect MovementManager's
-            (MovementManager, "way_point", "_mgr_way_point_unused"),
-        ]
-    )
-    .global_config(n_workers=8)
-)
