+++ dimos/robot/test_all_blueprints.py
@@ -29,14 +29,12 @@ OPTIONAL_ERROR_SUBSTRINGS = {
 # These need git LFS, so can't be run on the ubuntu runners.
 SELF_HOSTED_BLUEPRINTS = frozenset(
     {
-        "alfred-nav",
         "coordinator-basic",
         "coordinator-cartesian-ik-mock",
         "coordinator-cartesian-ik-piper",
         "coordinator-combined-xarm6",
         "coordinator-flowbase",
         "coordinator-flowbase-keyboard-teleop",
-        "coordinator-flowbase-nav",
         "coordinator-mobile-manip-mock",
         "coordinator-mock",
         "coordinator-mock-twist-base",
@@ -63,7 +61,6 @@ SELF_HOSTED_BLUEPRINTS = frozenset(
         "teleop-quest-xarm6",
         "teleop-quest-xarm7",
         "teleop-quest-xarm7-video",
-        "unitree-g1-nav-sim",
         "xarm-perception",
         "xarm-perception-agent",
         "xarm-perception-sim",
